Reset attr_params structure before setting default values
The attr_params structure can be reused multiple times for parsing
multiple m= blocks.
If for some value it gets updated during the parsing of an m= block but
the next m= block doesn't set this value, the previous value persists
and appears as set by the second m= block.
The solution is to reset the structure before every usage of it.

Signed-off-by: Ilan Smith <lunnys@gmail.com>