Unify format of extract_2110_xx_params() functions
In preparation for turning the extract_2110_xx_params() function into a
generic function which receives as a callback the specific attribute
assigning it is necessary to make all current such functions (for
SMPTE2110 -20, -30, -40) similar in structure.
Moved the connection details extracted in -20 only which are only
required for by extract_packet_info() to this function.

Signed-off-by: Ilan Smith <lunnys@gmail.com>