Accept an unspecified PM field in a=FMTP attributes
SMPTE ST2110-22 doesn't require the pm= field to be present in it's
a=fmtp attribute list. Due to the commit that sets the changes the
unspecified value to be default (0) the parsing was failing in this
case.o

Also fixed crash due to double freeing of group and media vectors as
well as the attributes list.

Signed-off-by: Ilan Smith <lunnys@gmail.com>