Prevent parsing FMTP attributes if media sub-type is unknown
The parsing of a=fmtp attributes depends on the particular media
subtype.
The subtype is typically set in the a=rtpmap attribute, hence the order
of attributes should be:

  a=rtpmap...
  a=fmtp...

If the a=rtpmap attribute does not precede the a=fmtp attribute the
parser will not know how to parse the specific a=fmtp parameters.

This commit causes the parser to exit elegantly in such a case.

Signed-off-by: Ilan Smith <lunnys@gmail.com>